protecting api routes, authorized request, protecting react routes

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,22 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
+
     async function handleClick() {
+        if(!user) {
+            return;
+        }
         const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await response.json();
         if (response.ok) dispatch({ type: 'DELETE_WORKOUT', payload: json });
